requirements.txt
pyserial==3.5
numpy==1.24.2